Move OpenCode serve to import-only module; remove dead wizard (PR 14, 16).
Extract findOpencodeBinary/fetchLocalProviders to opencode-serve.ts and delete
runWizard/askSubscriptionTier from the launch path.
